refactor: modulired code
