+++ handler.go
@@ -20,9 +20,12 @@ import (
 	"net/http"
 	"net/http/httputil"
 
+	"github.com/rond-authz/rond/core"
 	"github.com/rond-authz/rond/internal/config"
 	"github.com/rond-authz/rond/internal/mongoclient"
 	"github.com/rond-authz/rond/internal/opatranslator"
+	"github.com/rond-authz/rond/internal/utils"
+	"github.com/rond-authz/rond/openapi"
 
 	"github.com/mia-platform/glogger/v2"
 	"github.com/sirupsen/logrus"
@@ -38,8 +41,8 @@ func ReverseProxyOrResponse(
 	env config.EnvironmentVariables,
 	w http.ResponseWriter,
 	req *http.Request,
-	permission *RondConfig,
-	partialResultsEvaluators PartialResultsEvaluators,
+	permission *openapi.RondConfig,
+	partialResultsEvaluators core.PartialResultsEvaluators,
 ) {
 	if env.Standalone {
 		if permission.RequestFlow.GenerateQuery {
@@ -70,13 +73,13 @@ func rbacHandler(w http.ResponseWriter, req *http.Request) {
 		return
 	}
 
-	permission, err := GetXPermission(requestContext)
+	permission, err := openapi.GetXPermission(requestContext)
 	if err != nil {
 		logger.WithField("error", logrus.Fields{"message": err.Error()}).Error("no policy permission found in context")
 		failResponse(w, "no policy permission found in context", GENERIC_BUSINESS_ERROR_MESSAGE)
 		return
 	}
-	partialResultEvaluators, err := GetPartialResultsEvaluators(requestContext)
+	partialResultEvaluators, err := core.GetPartialResultsEvaluators(requestContext)
 	if err != nil {
 		logger.WithField("error", logrus.Fields{"message": err.Error()}).Error("no partialResult evaluators found in context")
 		failResponse(w, "no partialResult evaluators found in context", GENERIC_BUSINESS_ERROR_MESSAGE)
@@ -89,7 +92,13 @@ func rbacHandler(w http.ResponseWriter, req *http.Request) {
 	ReverseProxyOrResponse(logger, env, w, req, permission, partialResultEvaluators)
 }
 
-func EvaluateRequest(req *http.Request, env config.EnvironmentVariables, w http.ResponseWriter, partialResultsEvaluators PartialResultsEvaluators, permission *RondConfig) error {
+func EvaluateRequest(
+	req *http.Request,
+	env config.EnvironmentVariables,
+	w http.ResponseWriter,
+	partialResultsEvaluators core.PartialResultsEvaluators,
+	permission *openapi.RondConfig,
+) error {
 	requestContext := req.Context()
 	logger := glogger.Get(requestContext)
 
@@ -100,14 +109,14 @@ func EvaluateRequest(req *http.Request, env config.EnvironmentVariables, w http.
 		return err
 	}
 
-	input, err := createRegoQueryInput(req, env, permission.Options.EnableResourcePermissionsMapOptimization, userInfo, nil)
+	input, err := core.CreateRegoQueryInput(req, env, permission.Options.EnableResourcePermissionsMapOptimization, userInfo, nil)
 	if err != nil {
 		logger.WithField("error", logrus.Fields{"message": err.Error()}).Error("failed rego query input creation")
 		failResponseWithCode(w, http.StatusInternalServerError, "RBAC input creation failed", GENERIC_BUSINESS_ERROR_MESSAGE)
 		return err
 	}
 
-	var evaluatorAllowPolicy *OPAEvaluator
+	var evaluatorAllowPolicy *core.OPAEvaluator
 	if !permission.RequestFlow.GenerateQuery {
 		evaluatorAllowPolicy, err = partialResultsEvaluators.GetEvaluatorFromPolicy(requestContext, permission.RequestFlow.PolicyName, input, env)
 		if err != nil {
@@ -116,7 +125,7 @@ func EvaluateRequest(req *http.Request, env config.EnvironmentVariables, w http.
 			return err
 		}
 	} else {
-		evaluatorAllowPolicy, err = createQueryEvaluator(requestContext, logger, req, env, permission.RequestFlow.PolicyName, input, nil)
+		evaluatorAllowPolicy, err = core.CreateQueryEvaluator(requestContext, logger, req, env, permission.RequestFlow.PolicyName, input, nil)
 		if err != nil {
 			logger.WithField("error", logrus.Fields{"message": err.Error()}).Error("cannot create evaluator")
 			failResponseWithCode(w, http.StatusForbidden, "RBAC policy evaluator creation failed", NO_PERMISSIONS_ERROR_MESSAGE)
@@ -126,8 +135,8 @@ func EvaluateRequest(req *http.Request, env config.EnvironmentVariables, w http.
 
 	_, query, err := evaluatorAllowPolicy.PolicyEvaluation(logger, permission)
 	if err != nil {
-		if errors.Is(err, opatranslator.ErrEmptyQuery) && hasApplicationJSONContentType(req.Header) {
-			w.Header().Set(ContentTypeHeaderKey, JSONContentTypeHeader)
+		if errors.Is(err, opatranslator.ErrEmptyQuery) && utils.HasApplicationJSONContentType(req.Header) {
+			w.Header().Set(utils.ContentTypeHeaderKey, utils.JSONContentTypeHeader)
 			w.WriteHeader(http.StatusOK)
 			if _, err := w.Write([]byte("[]")); err != nil {
 				logger.WithField("error", logrus.Fields{"message": err.Error()}).Warn("failed response write")
@@ -163,7 +172,14 @@ func EvaluateRequest(req *http.Request, env config.EnvironmentVariables, w http.
 	return nil
 }
 
-func ReverseProxy(logger *logrus.Entry, env config.EnvironmentVariables, w http.ResponseWriter, req *http.Request, permission *RondConfig, partialResultsEvaluators PartialResultsEvaluators) {
+func ReverseProxy(
+	logger *logrus.Entry,
+	env config.EnvironmentVariables,
+	w http.ResponseWriter,
+	req *http.Request,
+	permission *openapi.RondConfig,
+	partialResultsEvaluators core.PartialResultsEvaluators,
+) {
 	targetHostFromEnv := env.TargetServiceHost
 	proxy := httputil.ReverseProxy{
 		FlushInterval: -1,
