+++ internal/utils/http_test.go
@@ -1,0 +1,50 @@
+package utils
+
+import (
+	"encoding/json"
+	"net/http"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestUnmarshalHeader(t *testing.T) {
+	userPropertiesHeaderKey := "miauserproperties"
+	mockedUserProperties := map[string]interface{}{
+		"my":  "other",
+		"key": []string{"is", "not"},
+	}
+	mockedUserPropertiesStringified, err := json.Marshal(mockedUserProperties)
+	require.NoError(t, err)
+
+	t.Run("header not exists", func(t *testing.T) {
+		headers := http.Header{}
+		var userProperties map[string]interface{}
+
+		ok, err := UnmarshalHeader(headers, userPropertiesHeaderKey, &userProperties)
+
+		require.True(t, !ok, "Unmarshal not existing header")
+		require.NoError(t, err, "Unexpected error if doesn't exist header")
+	})
+
+	t.Run("header exists but the unmarshalling fails", func(t *testing.T) {
+		headers := http.Header{}
+		headers.Set(userPropertiesHeaderKey, string(mockedUserPropertiesStringified))
+		var userProperties string
+
+		ok, err := UnmarshalHeader(headers, userPropertiesHeaderKey, &userProperties)
+		require.False(t, ok, "Unexpected success during unmarshalling")
+		var unmarshalErr = &json.UnmarshalTypeError{}
+		require.ErrorAs(t, err, &unmarshalErr, "Unexpected error on unmarshalling")
+	})
+
+	t.Run("header exists and unmarshalling finishes correctly", func(t *testing.T) {
+		headers := http.Header{}
+		headers.Set(userPropertiesHeaderKey, string(mockedUserPropertiesStringified))
+		var userProperties map[string]interface{}
+
+		ok, err := UnmarshalHeader(headers, userPropertiesHeaderKey, &userProperties)
+		require.True(t, ok, "Unexpected failure")
+		require.NoError(t, err, "Unexpected error")
+	})
+}
