+++ router.go
@@ -15,18 +15,22 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 	"net/http"
 	"path"
-	"regexp"
 	"sort"
 	"strings"
 
 	swagger "github.com/davidebianchi/gswagger"
+	"github.com/mia-platform/glogger/v2"
+	"github.com/rond-authz/rond/core"
 	"github.com/rond-authz/rond/internal/config"
 	"github.com/rond-authz/rond/internal/metrics"
 	"github.com/rond-authz/rond/internal/utils"
+	"github.com/rond-authz/rond/openapi"
 	"github.com/rond-authz/rond/types"
+	"github.com/sirupsen/logrus"
 
 	"github.com/gorilla/mux"
 )
@@ -103,7 +107,7 @@ func addStandaloneRoutes(router *swagger.Router) error {
 	return nil
 }
 
-func setupRoutes(router *mux.Router, oas *OpenAPISpec, env config.EnvironmentVariables) {
+func setupRoutes(router *mux.Router, oas *openapi.OpenAPISpec, env config.EnvironmentVariables) {
 	var documentationPermission string
 	documentationPathInOAS := oas.Paths[env.TargetServiceOASPath]
 	if documentationPathInOAS != nil {
@@ -120,8 +124,8 @@ func setupRoutes(router *mux.Router, oas *OpenAPISpec, env config.EnvironmentVar
 	for path, pathMethods := range oas.Paths {
 		paths = append(paths, path)
 		for method := range pathMethods {
-			if method == AllHTTPMethod {
-				methods[path] = OasSupportedHTTPMethods
+			if method == openapi.AllHTTPMethod {
+				methods[path] = openapi.OasSupportedHTTPMethods
 				continue
 			}
 			if methods[path] == nil {
@@ -143,17 +147,17 @@ func setupRoutes(router *mux.Router, oas *OpenAPISpec, env config.EnvironmentVar
 		}
 		if strings.Contains(pathToRegister, "*") {
 			pathWithoutAsterisk := strings.ReplaceAll(pathToRegister, "*", "")
-			router.PathPrefix(convertPathVariablesToBrackets(pathWithoutAsterisk)).HandlerFunc(rbacHandler).Methods(methods[path]...)
+			router.PathPrefix(openapi.ConvertPathVariablesToBrackets(pathWithoutAsterisk)).HandlerFunc(rbacHandler).Methods(methods[path]...)
 			continue
 		}
 		if path == env.TargetServiceOASPath && documentationPermission == "" {
-			router.HandleFunc(convertPathVariablesToBrackets(pathToRegister), alwaysProxyHandler).Methods(http.MethodGet)
+			router.HandleFunc(openapi.ConvertPathVariablesToBrackets(pathToRegister), alwaysProxyHandler).Methods(http.MethodGet)
 			continue
 		}
-		router.HandleFunc(convertPathVariablesToBrackets(pathToRegister), rbacHandler).Methods(methods[path]...)
+		router.HandleFunc(openapi.ConvertPathVariablesToBrackets(pathToRegister), rbacHandler).Methods(methods[path]...)
 	}
 	if documentationPathInOAS == nil {
-		router.HandleFunc(convertPathVariablesToBrackets(env.TargetServiceOASPath), alwaysProxyHandler)
+		router.HandleFunc(openapi.ConvertPathVariablesToBrackets(env.TargetServiceOASPath), alwaysProxyHandler)
 	}
 	// FIXME: All the routes don't inserted above are anyway handled by rbacHandler.
 	//        Maybe the code above can be cleaned.
@@ -166,14 +170,67 @@ func setupRoutes(router *mux.Router, oas *OpenAPISpec, env config.EnvironmentVar
 	router.PathPrefix(fallbackRoute).HandlerFunc(rbacHandler)
 }
 
-var matchColons = regexp.MustCompile(`\/:(\w+)`)
+func OPAMiddleware(opaModuleConfig *core.OPAModuleConfig, openAPISpec *openapi.OpenAPISpec, envs *config.EnvironmentVariables, policyEvaluators core.PartialResultsEvaluators) mux.MiddlewareFunc {
+	OASrouter := openAPISpec.PrepareOASRouter()
 
-func convertPathVariablesToBrackets(path string) string {
-	return matchColons.ReplaceAllString(path, "/{$1}")
-}
+	return func(next http.Handler) http.Handler {
+		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			if utils.Contains(routesToNotProxy, r.URL.RequestURI()) {
+				next.ServeHTTP(w, r)
+				return
+			}
 
-var matchBrackets = regexp.MustCompile(`\/{(\w+)}`)
+			path := r.URL.EscapedPath()
+			if envs.Standalone {
+				path = strings.Replace(r.URL.EscapedPath(), envs.PathPrefixStandalone, "", 1)
+			}
 
-func convertPathVariablesToColons(path string) string {
-	return matchBrackets.ReplaceAllString(path, "/:$1")
+			logger := glogger.Get(r.Context())
+
+			permission, err := openAPISpec.FindPermission(OASrouter, path, r.Method)
+			if r.Method == http.MethodGet && r.URL.Path == envs.TargetServiceOASPath && permission.RequestFlow.PolicyName == "" {
+				fields := logrus.Fields{}
+				if err != nil {
+					fields["error"] = logrus.Fields{"message": err.Error()}
+				}
+				logger.WithFields(fields).Info("Proxying call to OAS Path even with no permission")
+				next.ServeHTTP(w, r)
+				return
+			}
+
+			if err != nil || permission.RequestFlow.PolicyName == "" {
+				errorMessage := "User is not allowed to request the API"
+				statusCode := http.StatusForbidden
+				fields := logrus.Fields{
+					"originalRequestPath": utils.SanitizeString(r.URL.Path),
+					"method":              utils.SanitizeString(r.Method),
+					"allowPermission":     utils.SanitizeString(permission.RequestFlow.PolicyName),
+				}
+				technicalError := ""
+				if err != nil {
+					technicalError = err.Error()
+					fields["error"] = logrus.Fields{"message": err.Error()}
+					errorMessage = "The request doesn't match any known API"
+				}
+				if errors.Is(err, openapi.ErrNotFoundOASDefinition) {
+					statusCode = http.StatusNotFound
+				}
+				logger.WithFields(fields).Errorf(errorMessage)
+				failResponseWithCode(w, statusCode, technicalError, errorMessage)
+				return
+			}
+
+			ctx := openapi.WithXPermission(
+				core.WithOPAModuleConfig(
+					core.WithPartialResultsEvaluators(
+						openapi.WithRouterInfo(logger, r.Context(), r),
+						policyEvaluators,
+					),
+					opaModuleConfig,
+				),
+				&permission,
+			)
+			next.ServeHTTP(w, r.WithContext(ctx))
+		})
+	}
 }
