+++ utilities.go
@@ -17,24 +17,17 @@ package main
 import (
 	"encoding/json"
 	"net/http"
-	"strings"
 
+	"github.com/rond-authz/rond/internal/utils"
 	"github.com/rond-authz/rond/types"
 )
-
-const ContentTypeHeaderKey = "content-type"
-const JSONContentTypeHeader = "application/json"
-
-func hasApplicationJSONContentType(headers http.Header) bool {
-	return strings.HasPrefix(headers.Get(ContentTypeHeaderKey), JSONContentTypeHeader)
-}
 
 func failResponse(w http.ResponseWriter, technicalError, businessError string) {
 	failResponseWithCode(w, http.StatusInternalServerError, technicalError, businessError)
 }
 
 func failResponseWithCode(w http.ResponseWriter, statusCode int, technicalError, businessError string) {
-	w.Header().Set(ContentTypeHeaderKey, JSONContentTypeHeader)
+	w.Header().Set(utils.ContentTypeHeaderKey, utils.JSONContentTypeHeader)
 	w.WriteHeader(statusCode)
 	content, err := json.Marshal(types.RequestError{
 		StatusCode: statusCode,
@@ -47,13 +40,4 @@ func failResponseWithCode(w http.ResponseWriter, statusCode int, technicalError,
 
 	//#nosec G104 -- Intended to avoid disruptive code changes
 	w.Write(content)
-}
-
-func unmarshalHeader(headers http.Header, headerKey string, v interface{}) (bool, error) {
-	headerValueStringified := headers.Get(headerKey)
-	if headerValueStringified != "" {
-		err := json.Unmarshal([]byte(headerValueStringified), &v)
-		return err == nil, err
-	}
-	return false, nil
 }
