+++ utilities_test.go
@@ -21,51 +21,11 @@ import (
 	"net/http/httptest"
 	"testing"
 
+	"github.com/rond-authz/rond/internal/utils"
 	"github.com/rond-authz/rond/types"
 
 	"github.com/stretchr/testify/require"
 )
-
-func TestUnmarshalHeader(t *testing.T) {
-	userPropertiesHeaderKey := "miauserproperties"
-	mockedUserProperties := map[string]interface{}{
-		"my":  "other",
-		"key": []string{"is", "not"},
-	}
-	mockedUserPropertiesStringified, err := json.Marshal(mockedUserProperties)
-	require.NoError(t, err)
-
-	t.Run("header not exists", func(t *testing.T) {
-		headers := http.Header{}
-		var userProperties map[string]interface{}
-
-		ok, err := unmarshalHeader(headers, userPropertiesHeaderKey, &userProperties)
-
-		require.True(t, !ok, "Unmarshal not existing header")
-		require.NoError(t, err, "Unexpected error if doesn't exist header")
-	})
-
-	t.Run("header exists but the unmarshalling fails", func(t *testing.T) {
-		headers := http.Header{}
-		headers.Set(userPropertiesHeaderKey, string(mockedUserPropertiesStringified))
-		var userProperties string
-
-		ok, err := unmarshalHeader(headers, userPropertiesHeaderKey, &userProperties)
-		require.False(t, ok, "Unexpected success during unmarshalling")
-		var unmarshalErr = &json.UnmarshalTypeError{}
-		require.ErrorAs(t, err, &unmarshalErr, "Unexpected error on unmarshalling")
-	})
-
-	t.Run("header exists and unmarshalling finishes correctly", func(t *testing.T) {
-		headers := http.Header{}
-		headers.Set(userPropertiesHeaderKey, string(mockedUserPropertiesStringified))
-		var userProperties map[string]interface{}
-
-		ok, err := unmarshalHeader(headers, userPropertiesHeaderKey, &userProperties)
-		require.True(t, ok, "Unexpected failure")
-		require.NoError(t, err, "Unexpected error")
-	})
-}
 
 func TestFailResponseWithCode(t *testing.T) {
 	w := httptest.NewRecorder()
@@ -73,7 +33,7 @@ func TestFailResponseWithCode(t *testing.T) {
 	failResponseWithCode(w, http.StatusInternalServerError, "The Error", "The Message")
 	require.Equal(t, http.StatusInternalServerError, w.Result().StatusCode)
 
-	require.Equal(t, JSONContentTypeHeader, w.Result().Header.Get(ContentTypeHeaderKey))
+	require.Equal(t, utils.JSONContentTypeHeader, w.Result().Header.Get(utils.ContentTypeHeaderKey))
 
 	bodyBytes, err := io.ReadAll(w.Body)
 	require.NoError(t, err)
